Refactor - rename and reorganize
